+++ src/components/fileLoaders/VisualUploadFile/VisualUploadFile.tsx
@@ -31,11 +31,11 @@ const VisualUploadFile = ({
       )}
     >
       {selectedImage && (
-        <div className={"upload-container"}>
+        <div className="upload-container">
           <img
             alt="not fount"
             src={URL.createObjectURL(selectedImage)}
-            className={"uplaoded-img"}
+            className="uploaded-img"
           />
         </div>
       )}
@@ -43,7 +43,7 @@ const VisualUploadFile = ({
       <input
         id="load-file"
         type="file"
-        className={"load-file__input"}
+        className="load-file__input"
         onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
           setSelectedImage(event?.target?.files && event?.target?.files[0]);
         }}
@@ -58,11 +58,7 @@ const VisualUploadFile = ({
             {title}
           </label>
           {backgroundImg && (
-            <img
-              src={backgroundImg}
-              alt=""
-              className={"load-file__background"}
-            />
+            <img src={backgroundImg} alt="" className="load-file__background" />
           )}
         </div>
       )}