+++ src/components/inputs/Select/Select.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import cn from "classnames";
 import Select from "react-select";
 
@@ -28,18 +28,46 @@ const SelectInput = ({
   styles,
   modalPortalTarget = document.body
 }: SelectInputInterface) => {
+  const [hasValue, setHasValue] = useState(false);
+  const [isFocused, setIsFocused] = useState(false);
+
+  const handleChange = (e: any) => {
+    e.value && setHasValue(true);
+    onChange && onChange(e);
+  };
+
+  const handleFocus = () => {
+    setIsFocused(true);
+  };
+
+  const handleBlur = () => {
+    !hasValue && setIsFocused(false);
+    onBlur && onBlur();
+  };
+
   return (
-    <Select
-      options={options}
-      className={cn("select-container", className)}
-      classNamePrefix="select"
-      placeholder={placeholder}
-      name={name}
-      onChange={onChange}
-      onBlur={onBlur}
-      styles={styles}
-      menuPortalTarget={modalPortalTarget}
-    />
+    <div className="select-wrapper">
+      <Select
+        options={options}
+        className={cn("select-container", className)}
+        classNamePrefix="select"
+        placeholder={placeholder}
+        name={name}
+        onChange={handleChange}
+        onBlur={handleBlur}
+        onFocus={handleFocus}
+        styles={styles}
+        menuPortalTarget={modalPortalTarget}
+        defaultMenuIsOpen={true}
+      />
+      <p
+        className={cn("select-wrapper__placeholder", {
+          "select-wrapper__placeholder_active": isFocused
+        })}
+      >
+        {placeholder}
+      </p>
+    </div>
   );
 };