+++ Changelog.md
@@ -1,4 +1,5 @@
-# gov-ua-ui changelog from v0.3.4. to v0.3.5.
+# gov-ua-ui changelog from v0.3.5. to v0.3.6.
 
-1. Refactored <Dialog> component, added possibility to pass custom controls.
-2. Added onClick handler to Checkbox component.
+1. Reduced size of <VisualUploadFile/> component.
+2. Fixed <Select/> label styles when component active. 
+3. Fixed keys in <Navigation />.